+++ pos_multi_session/models/pos_multi_session_models.py
@@ -21,7 +21,7 @@ class PosConfig(models.Model):
                                             'Deselect checkbox of "Active" if this POS should not use syncing.'
                                             'Before updating it you need to close active session',
                                        default=lambda self: self.env.ref('pos_multi_session.default_multi_session', raise_if_not_found=False))
-    multi_session_active = fields.Boolean(string="Active", help="Select the checkbox to enable synchronization for this POS", default=False)
+    multi_session_active = fields.Boolean(string="Active", help="Select the checkbox to enable synchronization for this POS", default=True)
     multi_session_accept_incoming_orders = fields.Boolean('Accept incoming orders', default=True)
     multi_session_replace_empty_order = fields.Boolean('Replace empty order', default=True, help='Empty order is deleted whenever new order is come from another POS')
     multi_session_deactivate_empty_order = fields.Boolean('Deactivate empty order', default=False, help='POS is switched to new foreign Order, if current order is empty')
@@ -57,13 +57,16 @@ class PosMultiSession(models.Model):
     fiscal_position_ids = fields.Many2many('account.fiscal.position', string='Fiscal Positions', ondelete="restrict")
 
     @api.multi
-    def action_default_multi_session(self):
+    def action_set_default_multi_session(self):
+        """
+            during installation of the module set default multi-session
+            for all POSes for which not specified multi_session_id
+        """
         self.ensure_one()
         configs = self.env['pos.config'].search([('multi_session_id', '=', False)])
-        for pos in configs:
-            pos.write({
-                'multi_session_id': self.id,
-            })
+        configs.write({
+            'multi_session_id': self.id
+        })
 
 
 class PosSession(models.Model):