+++ pos_multi_session/doc/changelog.rst
@@ -1,6 +1,6 @@
 `4.1.0`
 -------
-- **NEW:** Field "Active" to enable/disable synchronization for POSes. The multi-session parameter is required now to provide compatibility with "POS Multi Session Menu" module
+- **NEW:** Field "Active" to enable/disable synchronization for POSes. The multi-session parameter is required now to provide compatibility with "POS Multi Session Menu" module and other modules that put common settings in "POS Multi Session"
 
 `4.0.6`
 -------